[178622 Patch Set 2] Show runtime permission dialog in order to read storage
In Marshmallow, HTMLViewer can not open text file
because READ_EXTERNAL_STORAGE permission is off by default.
In addition, it just shows up 'Webpage not available', so users
can't notice that this is the permission problem. It would be better
to show dialog to check permission for read storage in HTMLViewer.

Change-Id: I26ad05136194cb36d4a17d3a8ab1a3fd2b5dfeb5

Signed-off-by: amoseui <amoseui@gmail.com>